part6: ex6.8
